Fix merge point calculation for repeated merges with same branch commit
A commit might be ancestor of multiple merges into the mainline. Only the
first of those merges is considered the merge point of that commit. To find
it we have to traverse mainline merges from oldest to newest, storing the
first appearance of a commit as ancestor of a merge.