Override end-of-line attributes in gitattributes files
If a repo contains .gitattributes files and contains special attributes
for line endings, they are not respected while fetching using go-git
and default configuration values are followed. This leads to a dirty
workspace.

To avoid this, we need to instruct git to not attempt any line ending
conversions. The only way to override the config in .gitattributes
files is through a .git/info/attributes file.

This commit adds such a file and prevents any line ending conversions.