Don't apply empty patch when merge commit is direct descendant of branch point commit
When f_latest_merge_commit is a direct descendant of
f_latest_branch_point_commit, there will be no diff between them.

Applying an empty patch will give an error. So, apply the patch only if
it exists and allow an empty commit so that the go.mod changes can be squashed
with this commit later, if needed.