sync-tags: add gomodules support
This commit adds gomodules support for tagging (to sync-tags) to make sure
that kube dependencies point to the commit sha at the relevant tag in go.mod.

If the repo at the tag has a go.mod file, sync-tags pins the dependencies
at the pseudoversion pointing to the tag and packages up into the local
go mod cache, so that it can be downloaded and used by `go mod tidy`.
After that, godeps-gen is used to generate Godeps.json from the go.mod file.

If the repo at the tag does not contain a go.mod file, sync-tags directly
updates Godeps.json to point to the commit sha at the tag (old behaviour).