Add gomod-zip to correctly create the zip file
The zip file created by go mod download has extra normalization over
the zip file created by git archive. The normalization process is done below.

While the normalization can also be achieved via a simple zip command, the zip file
created by go mod download has the `00-00-1980 00:00` timestamp in the file header
for all files in the zip archive. This is not a valid UNIX timestamp and cannot be
set easily. This is, however, valid in MSDOS. The `archive/zip` package uses the
MSDOS version so we create the zip file using this package.