Allow specifying branch-specific smoke tests
Currently, smoke tests are specified for the entire repo and apply to
all branches. However, this leads to problems when a single test cannot
be run across all branches.

For example, the smoke test for client-go uses `go test -mod=mod ./...`.
The `-mod=mod` option is available only from go1.14 and is necessary for
running tests with go1.16.

client-go `release-1.18` branch uses go1.13.15 which doesn't support the
`-mod=mod` option, so the tests fail.

This commit fixes this issue by additonally allowing to specify smoke
tests per branch. The current option for tests that apply to all
branches is not changed to maintain compatibility.