util.sh: avoid obsolete go mod download before packaging up repo in cache
Without this commit, if a wrong hash has already been published in
go.sum, we get the following error (from https://github.com/kubernetes/kubernetes/issues/56876#issuecomment-828176255):

```
verifying k8s.io/client-go@v0.0.0-20210401152202-307e3a38a167: checksum mismatch
    downloaded: h1:opqPJ8l/hqiKIseiNVxcUiYF0vc8gAcyy/rZZrxseTs=
    go.sum:     h1:6dfdaZ0/RfxMbyUuuKL4xdk4+DVjkO76nmb6z4S+bn0=

SECURITY ERROR
This download does NOT match an earlier download recorded in go.sum.
The bits may have been replaced on the origin server, or an attacker may
have intercepted the download attempt.
```

In this case, `cli-runtime` contains an incorrect value of the hash in
it's go.sum because the packaging was broken and incorrect zip files
were published. Now that the packaging has been fixed i.e a correctly
hashed zip file has been created in the go mod cache, the `downloaded`
hash represents the correct value.

However, it's `go.sum` still contains the old incorrect value which
causes `go mod download` to complain with a checksum error.

Since we don't care about the dependent repo's dependencies, we can
remove the `go mod download` step here. The go.sum will be fixed in
`update-deps-in-gomod()`.

More details on why it is safe to remove `go mod download` here:

If a repo contains other staging repos as dependencies, we do this:

1) Checkout out the dependent staging repo at the desired commit

2) Run `go mod download` to add the _dependencies of the dependent
staging repo_ to the go mod cache.

3) Since `go mod download` can bring in changes to `go.sum` for the
_dependencies of the dependent staging repo_, checkout `go.sum` to the
latest HEAD. We don't care about changes to go.sum of the dependent
repo (we only care about go.mod since we need to package it up).

4) Check if the dependent repo already exists in the go mod cache
at the desired commit. If yes, do nothing.

5) If no, create go.mod, zip and info files in the go mod cache.

In 2 and 3, we download dependencies we don't care about and also checkout
`go.sum` to it's original value. So these lines are safe to remove and
do not affect the packaging up of the dependent repo.