util.sh: remove go.sum only if it is not empty
There may be cases when a repo's go.sum is an empty file. For example:
`release-1.19` branch of https://github.com/kubernetes/controller-manager.
This can happen if a repo was part of a release but did not include any
code when it was first added.

In such a case, the go.sum file for the repo exists but is empty:
https://github.com/kubernetes/controller-manager/blob/release-1.19/go.sum
because there are no dependencies.

If we remove the go.sum file for such a repo, it will NOT be recreated
in the next step when we run `go mod tidy`. So this commit only removes
the go.sum file if it is non-empty.

Note that we remove the go.sum file to fix incorrect hashes in go.sum.
Since go.sum is empty in this case, it doesn't matter if we don't
remove the go.sum file.