Choose only one dst branch point commit for specified k/k commits
Due to some changes in the release tooling, it is now possible to have
two k/k PRs/commits with the same k/k mainline commit. This causes one of the
k/k commit to not be published.

We manually published this missing commit in the staging repo.
Due to this, the bot now re-publishes commits created after the manually added commit.
If a new branch exists at one of the re-published commits, it leads to
two dst_branch_point_commits.

So, allow specifying a list of k/k commits such that only a single
dst_branch_point_commit is picked for them.

See slack thread for more details - https://kubernetes.slack.com/archives/C2C40FMNF/p1638139647251300