doc: update steps to cleanup PVC
Update steps to scale down the replica set and then delete the PVC.
This make sure that the old PVC does not get Bound to the new Pod,
when only the pod is deleted. After the the old PVC is deleted and
then recreated the replica can be scaled up.

Signed-off-by: Akhil Mohan <akhilerm@gmail.com>